E-Diaristas-Backend/requirements.txt
arrow==1.2.0
asgiref==3.4.1
beautifulsoup4==4.9.3
blessed==1.19.0
boto3==1.20.29
botocore==1.23.29
CacheControl==0.12.10
certifi==2021.5.30
cffi==1.15.0
chardet==3.0.4
charset-normalizer==2.0.3
coreapi==2.3.3
coreschema==0.0.4
cryptography==35.0.0
dj-database-url==0.5.0
Django==3.2.10
django-adminlte3==0.1.6
django-bootstrap4==3.0.1
django-cors-headers==3.7.0
django-environ==0.4.5
django-localflavor==3.1
django-picklefield==3.0.1
django-q==1.3.9
django-rest-passwordreset==1.2.1
django-storages==1.12.3
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
drf-yasg==1.20.0
googlemaps==4.5.3
gunicorn==20.1.0
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.1
jmespath==0.10.0
jsonpickle==0.9.6
MarkupSafe==2.0.1
msgpack==1.0.2
packaging==21.0
pagarme-python==4.0.1
Pillow==8.3.2
pycparser==2.21
PyJWT==2.1.0
pyOpenSSL==21.0.0
pyparsing==2.4.7
python-dateutil==2.8.2
python-stdnum==1.16
pytz==2021.1
redis==3.5.3
requests==2.23.0
ruamel.yaml==0.17.16
ruamel.yaml.clib==0.2.6
s3transfer==0.5.0
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.2
uritemplate==3.0.1
urllib3==1.25.9
wcwidth==0.2.5
whitenoise==5.3.0
